requirements.txt
RlGlue-andnp~=1.0
PyExpUtils-andnp~=6.0
PyFixedReps-andnp~=4.0
pyrlenvs-andnp~=2.0
PyExpPlotting-andnp
RlEvaluation~=1.0
numpy>=1.22.0
numba>=0.55.0

matplotlib
flake8
